MoonRulerLinesType으로 룰러타입 수정할 수 있도록 수정
