+++ lib/ruler_picker/element/moon_ruler_drawer.dart
@@ -58,7 +58,7 @@ class _MoonRulerDrawer {
       case MoonRulerLinesType.line:
         return _RulerLines(selectedNumber: selectedNumber, color: color, selectedColor: selectedColor, strokeWidth: strokeWidth, rulerPickerHeight: rulerPickerHeight, size: size, lineBetweenAlignWidth: lineBetweenAlignWidth, longVerticalLineHeightRatio: longVerticalLineHeightRatio, shortVerticalLineHeightRatio: shortVerticalLineHeightRatio, maxNumber: maxNumber, minNumber: minNumber, pickedVerticalLineHeightRatio: pickedVerticalLineHeightRatio, textStyle: textStyle, textPainter: textPainter);
       case MoonRulerLinesType.lineWithLabel:
-        return _RulerCircularLines(selectedNumber: selectedNumber, color: color, selectedColor: selectedColor, strokeWidth: strokeWidth, rulerPickerHeight: rulerPickerHeight, size: size, lineBetweenAlignWidth: lineBetweenAlignWidth, longVerticalLineHeightRatio: longVerticalLineHeightRatio, shortVerticalLineHeightRatio: shortVerticalLineHeightRatio, maxNumber: maxNumber, minNumber: minNumber, pickedVerticalLineHeightRatio: pickedVerticalLineHeightRatio, textStyle: textStyle, textPainter: textPainter);
+        return _RulerLinesWithLabel(selectedNumber: selectedNumber, color: color, selectedColor: selectedColor, strokeWidth: strokeWidth, rulerPickerHeight: rulerPickerHeight, size: size, lineBetweenAlignWidth: lineBetweenAlignWidth, longVerticalLineHeightRatio: longVerticalLineHeightRatio, shortVerticalLineHeightRatio: shortVerticalLineHeightRatio, maxNumber: maxNumber, minNumber: minNumber, pickedVerticalLineHeightRatio: pickedVerticalLineHeightRatio, textStyle: textStyle, textPainter: textPainter);
       case MoonRulerLinesType.circularLine:
         return _RulerCircularLines(selectedNumber: selectedNumber, color: color, selectedColor: selectedColor, strokeWidth: strokeWidth, rulerPickerHeight: rulerPickerHeight, size: size, lineBetweenAlignWidth: lineBetweenAlignWidth, longVerticalLineHeightRatio: longVerticalLineHeightRatio, shortVerticalLineHeightRatio: shortVerticalLineHeightRatio, maxNumber: maxNumber, minNumber: minNumber, pickedVerticalLineHeightRatio: pickedVerticalLineHeightRatio, textStyle: textStyle, textPainter: textPainter);
     }
