+++ lib/ruler_picker_lib.dart
@@ -12,10 +12,13 @@ import 'package:flutter/services.dart';
 part 'ruler_picker/ruler_picker.dart';
 
 part 'ruler_picker/element/moon_ruler_drawer.dart';
+
 part 'ruler_picker/element/picked_ruler_line.dart';
+part 'ruler_picker/element/picked_circular_ruler_line.dart';
 
 part 'ruler_picker/element/ruler_lines.dart';
 part 'ruler_picker/element/ruler_lines_with_label.dart';
 part 'ruler_picker/element/ruler_circular_lines.dart';
+
 
 part 'ruler_picker/ui_model/lines_type.dart';
